+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## HEAD
+## 2.14.0 (2018-12-27)
 
 * Add `tempfile` plugin for easier reusing of the same uploaded file copy on disk (@janko-m)